+++ src/clj/r99c/routes/home.clj
@@ -12,6 +12,7 @@
    [r99c.layout :as layout]
    [r99c.middleware :as middleware]
    [ring.util.response :refer [redirect]]
+   [selmer.filters :refer [add-filter!]]
    [taoensso.timbre :as timbre]))
 
 (timbre/set-level! :debug)
@@ -41,6 +42,16 @@
   [col n]
   {:n n :stat (if (lazy-contains? col n) "solved" "yet")})
 
+;; FIXME: does not wrap japanese strings.
+(defn- wrap
+  "fold string s at column n"
+  [n s]
+  (if (< (count s) n)
+    s
+    (str (subs s 0 n) "\n" (wrap n (subs s n)))))
+
+(add-filter! :wrap66  (fn [x] (wrap 66 x)))
+
 (defn status-page
   "display user's status. how many problems he/she solved?"
   [request]
@@ -59,7 +70,8 @@
       :recents (db/recent-answers {:n 10})
       :comments (db/sent-comments {:login login})
       :individual-chart (individual-chart individual period 600 150)
-      :class-chart (class-chart all-answers period 600 150)})))
+      :class-chart (class-chart all-answers period 600 150)
+      :recent-comments (db/recent-comments {:n 15})})))
 
 (defn problems-page
   "display problems."
@@ -130,16 +142,13 @@
   [request]
   (let [id (Integer/parseInt (get-in request [:path-params :id]))
         answer (db/get-answer-by-id {:id id})
-        num (:num answer)
-        problem  (db/get-problem {:num num})
-        comments (db/get-comments {:a_id id})
-        same-md5 (db/answers-same-md5 {:md5 (:md5 answer)})]
+        num (:num answer)]
     (if (db/get-answer {:num num :login (login request)})
       (layout/render request "comment-form.html"
-                     {:problem problem
-                      :answer answer
-                      :comments comments
-                      :same-md5 same-md5})
+                     {:answer answer
+                      :problem  (db/get-problem {:num num})
+                      :same-md5 (db/answers-same-md5 {:md5 (:md5 answer)})
+                      :comments (db/get-comments {:a_id id})})
       (layout/render request "error.html"
                      {:status 403
                       :title "Access Forbidden"